requirements.txt
streamlit
pillow
imagehash
pyperclip
PySimpleGUI
numpy
opencv-python-headless
